# БР1.2/Колесников Игорь Евгеньевич/homeworks/hw4/openapi.yaml
openapi: 3.0.3
info:
  title: Jobby — Microservices API
  description: |
    REST API для сайта по поиску работы, реализованный в виде микросервисной архитектуры.

    ## Сервисы
    | Сервис | Порт | Описание |
    |---|---|---|
    | user-service | 3101 | Аутентификация, регистрация, профили |
    | vacancy-service | 3102 | Компании, вакансии |
    | resume-service | 3103 | Резюме соискателей |
    | application-service | 3104 | Отклики на вакансии |
    | dictionary-service | 3105 | Справочники (read-only) |

    ## Аутентификация
    JWT хранится в `httpOnly` cookie (`access_token`). Каждый сервис самостоятельно верифицирует токен через `process.env.JWT_SECRET`.

    ## Внутренние эндпоинты (inter-service)
    Эндпоинты с префиксом `/internal/` предназначены для вызовов между сервисами внутри Docker-сети. В учебном Compose сервисные порты опубликованы на хосте, поэтому такие маршруты технически достижимы с хоста; доступ ограничен проверкой заголовка `X-Internal-Secret`, JWT для них не требуется.
  version: 2.0.0
servers:
  - url: http://localhost:3101
    description: user-service
  - url: http://localhost:3102
    description: vacancy-service
  - url: http://localhost:3103
    description: resume-service
  - url: http://localhost:3104
    description: application-service
  - url: http://localhost:3105
    description: dictionary-service
security:
  - cookieAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    internalSecret:
      type: apiKey
      in: header
      name: X-Internal-Secret
      description: Секрет для межсервисных запросов (из env INTERNAL_SECRET)
  responses:
    Unauthorized:
      description: Не передан или невалидный JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            message: Unauthorized
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 403
            message: Forbidden
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            message: Not found
    UnprocessableEntity:
      description: Ошибка валидации тела запроса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Bad request
    ValidationErrorResponse:
      type: object
      required: [statusCode, message, errors]
      properties:
        statusCode:
          type: integer
          example: 422
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # ── Auth / User ──
...
    CompanyShort:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
    CompanyDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        industry:
          type: object
          nullable: true
          allOf:
            - $ref: '#/components/schemas/IndustryItem'
        city:
          type: object
          nullable: true
          allOf:
            - $ref: '#/components/schemas/CityShort'
    CreateCompanyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: ООО Технологии
        description:
          type: string
        website:
          type: string
          example: https://example.com
        logoUrl:
          type: string
        industryId:
          type: string
          format: uuid
        cityId:
          type: string
          format: uuid

    # ── Vacancy ──
    VacancyShort:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        company:
          type: object
          nullable: true
          allOf:
            - $ref: '#/components/schemas/CompanyShort'
        city:
          type: object
          nullable: true
          allOf:
            - $ref: '#/components/schemas/CityShort'
        salaryType:
          type: string
          enum: [FIXED, RANGE, FROM, TO, NEGOTIABLE]
        salaryFixed:
          type: integer
          nullable: true
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        currency:
          type: string
          example: RUB
        isRemote:
          type: boolean
        experienceYearsMin:
          type: integer
          nullable: true
        updatedAt:
          type: string
          format: date-time
    VacancyDetail:
      allOf:
        - $ref: '#/components/schemas/VacancyShort'
        - type: object
          properties:
            description:
              type: string
              nullable: true
            conditions:
              type: string
              nullable: true
            requirementsNote:
              type: string
              nullable: true
            experienceYearsMax:
              type: integer
              nullable: true
            employmentType:
              type: object
              nullable: true
              allOf:
                - $ref: '#/components/schemas/EmploymentTypeItem'
            industry:
              type: object
              nullable: true
              allOf:
                - $ref: '#/components/schemas/IndustryItem'
            skills:
              type: array
              items:
                $ref: '#/components/schemas/VacancySkillItem'
            employer:
              $ref: '#/components/schemas/EmployerShort'
            createdAt:
              type: string
              format: date-time
    CreateVacancyRequest:
      type: object
      required: [title, salaryType, companyId]
      properties:
        title:
          type: string
          example: Backend Developer
        description:
          type: string
        conditions:
          type: string
        requirementsNote:
          type: string
        cityId:
          type: string
          format: uuid
        industryId:
          type: string
          format: uuid
        employmentTypeId:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        salaryType:
          type: string
          enum: [FIXED, RANGE, FROM, TO, NEGOTIABLE]
        salaryFixed:
          type: integer
          nullable: true
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        currency:
          type: string
          default: RUB
        experienceYearsMin:
          type: integer
          nullable: true
        experienceYearsMax:
          type: integer
          nullable: true
        isRemote:
          type: boolean
          default: false
        isPublished:
          type: boolean
          default: false
        skills:
          type: array
          items:
            type: object
            required: [skillId]
            properties:
              skillId:
                type: string
                format: uuid
    UpdateVacancyRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        conditions:
          type: string
        requirementsNote:
          type: string
        cityId:
          type: string
          format: uuid
        industryId:
          type: string
          format: uuid
        employmentTypeId:
          type: string
          format: uuid
        salaryType:
          type: string
          enum: [FIXED, RANGE, FROM, TO, NEGOTIABLE]
        salaryFixed:
          type: integer
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        currency:
          type: string
        experienceYearsMin:
          type: integer
        experienceYearsMax:
          type: integer
        isRemote:
          type: boolean
        isPublished:
          type: boolean
    VacancySkillItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        skill:
          $ref: '#/components/schemas/SkillShort'
...
    ApplicationItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vacancyId:
          type: string
          format: uuid
        resumeId:
          type: string
          format: uuid
        seekerUserId:
          type: string
          format: uuid
        coverLetter:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, VIEWED, INVITED, REJECTED, ACCEPTED]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CountryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Россия
    CityShort:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Санкт-Петербург
        country:
          $ref: '#/components/schemas/CountryItem'
    IndustryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Информационные технологии
    SkillShort:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: TypeScript
    EmploymentTypeItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: FULL_TIME
...
    EmployerShort:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        position:
          type: string
          nullable: true

# ─────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────
paths:
...
  /api/v1/companies:
    get:
      tags: [Companies]
      operationId: getCompanies
      summary: Список компаний (публично)
      security: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CompanyShort'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer

    post:
      tags: [Companies]
      operationId: createCompany
      summary: Создать компанию (только EMPLOYER)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        '201':
          description: Компания создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/companies/{companyId}:
    parameters:
      - name: companyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Companies]
      operationId: getCompany
      summary: Страница компании (публично)
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Companies]
      operationId: updateCompany
      summary: Обновить компанию (только сотрудник компании)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/vacancies:
    get:
      tags: [Vacancies]
      operationId: getVacancies
      summary: Список опубликованных вакансий (публично)
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: cityId
          in: query
          schema:
            type: string
            format: uuid
        - name: industryId
          in: query
          schema:
            type: string
            format: uuid
        - name: employmentTypeId
          in: query
          schema:
            type: string
            format: uuid
        - name: salaryMin
          in: query
          description: Минимально приемлемая зарплата
          schema:
            type: integer
            minimum: 0
        - name: experienceYearsMax
          in: query
          description: Максимальный стаж соискателя в годах
          schema:
            type: integer
            minimum: 0
        - name: skillIds
          in: query
          description: UUID навыков через запятую; вакансия должна содержать все перечисленные навыки
          schema:
            type: string
        - name: isRemote
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VacancyShort'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '400':
          description: Некорректные параметры фильтрации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Vacancies]
      operationId: createVacancy
      summary: Создать вакансию (только EMPLOYER)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVacancyRequest'
      responses:
        '201':
          description: Вакансия создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacancyDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Компания не найдена (vacancyService вызывает /internal/companies/:id)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 404
                message: Company not found
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/vacancies/{vacancyId}:
    parameters:
      - name: vacancyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Vacancies]
      operationId: getVacancy
      summary: Детали вакансии (публично)
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacancyDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Vacancies]
      operationId: updateVacancy
      summary: Обновить вакансию (только автор)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVacancyRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacancyDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Vacancies]
      operationId: deleteVacancy
      summary: Удалить вакансию (только автор)
      responses:
        '204':
          description: Удалено
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/employer/vacancies:
    get:
      tags: [Vacancies]
      operationId: getEmployerVacancies
      summary: Мои вакансии (только EMPLOYER)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VacancyShort'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/vacancies/{vacancyId}/applications:
    parameters:
      - name: vacancyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Vacancies]
      operationId: getVacancyApplications
      summary: Отклики на вакансию (только автор вакансии)
      description: vacancy-service проксирует запрос к application-service
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── VACANCY-SERVICE — Internal ──
...
  /internal/vacancies/{vacancyId}/applications:
    get:
      tags: [Applications]
      operationId: internalGetApplicationsByVacancy
      summary: Получить отклики на вакансию для vacancy-service
      security:
        - internalSecret: []
      parameters:
        - name: vacancyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Список откликов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationItem'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ══════════════════════════════════════
  # DICTIONARY-SERVICE — Public (server: :3105)
  # ══════════════════════════════════════
...